requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
django-environ
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
requests
